+++ dev-infra/release/BUILD.bazel
@@ -3,16 +3,13 @@ load("@npm_bazel_typescript//:index.bzl", "ts_library")
 ts_library(
     name = "release",
     srcs = glob([
-        "*.ts",
+        "**/*.ts",
     ]),
     module_name = "@angular/dev-infra-private/release",
     visibility = ["//dev-infra:__subpackages__"],
     deps = [
         "//dev-infra/utils",
-        "@npm//@types/node",
-        "@npm//@types/shelljs",
         "@npm//@types/yargs",
-        "@npm//shelljs",
         "@npm//yargs",
     ],
 )
